rage - add album art fetching for audio only stuff.
if the track is audio only, it'll look in a cache of albumart, if not,
request the big google in the sky for an image with some search terms
guessed from id3/metadata/filename and then use that. tak'es the first
thing found. not very smart, but pretty damned good most of the time.