Delete current media from playlist
Summary:
So you have unorganized files, or pseudo-organized files, and you dump them into rage. A song or video comes up that you don't want to bother with, so you want to hit delete to not play it.

This will be more useful later with a "loop all" feature, but I implemented it now because I often bump into this situation.

Reviewers: etrunko, raster

Projects: #rage

Differential Revision: https://phab.enlightenment.org/D1486

* also fix mem leak in not freeing vid struct content