rage browser mode and ui - slicken between desktop and fullscreen mode
so now sizing and count and spacing of items change, header
sizes/style change and more when switching to/from fullscreen mode.
fullscreen == "i want to be a media cener now on a big screen tv" vs
"i'm a video collection browser in a window on your desktop". far
nicer now. i think this is getting close to another release. maybe
after 1.18