browser videos dir - if efreet reports a silly dir for videos... handle
so if someone is sily and sets their videos dir to be $HOME ... don't
use it as frankly that makes the video browser useless. it is not
meant for such a mess of dirs. it's meant for a special subdirectory
system which is primarly just media files and nothing else ... not $HOME