rage videothumb - do not spawn multiple thumbnailers for the same file
so if you slide up and down the seek bar without an existing thumb
file (al my videos had thumb fils because i've been using the browser
which generates the thumb files for me mostly), then rage may spawn
several thumbnailersfor the same file. this is not a good thing. so
avoid it and only have 1 thumbnailer per file