browser - have a nicer view layout that doesnt expand items fully
so resize rage up and down and well... kind of undesireable things
happen to browser items in windowed mode. you want more of a swatch of
things to select from that don't get sily big (or way to horizontally
or vertically squashed) - similarly for fullscreen mode but be a
little different and don't enforce a max size then... it just works
better now... resizing around... try and enjoy. now back to work...